// apps/web/src/app/api/model/meSettingsResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Standup API
 * API do serviço de geração, revisão e publicação de standups.
 * OpenAPI spec version: 0.0.1
 */
import type { MeSettingsRecordDto } from './meSettingsRecordDto';

export interface MeSettingsResponseDto {
  data: MeSettingsRecordDto;
}